+++ src/controllers/cardController.js
@@ -20,29 +20,6 @@ export const createCard = async (req, res) => {
   }
 };
 
-export const getCardsForColumn = async (req, res) => {
-  try {
-    const errors = validationResult(req);
-
-    if (!errors.isEmpty()) {
-      return res.status(400).json(errors.array());
-    }
-
-    const cards = await ColumnCardModel.find({ boardId: req.body.boardId });
-    const columnCards = cards.filter(
-      (cards) => cards.columnId === req.body.columnId
-    );
-    console.log(cards);
-
-    res.json(columnCards);
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({
-      message: "Card getting error",
-    });
-  }
-};
-
 export const updateCard = async (req, res) => {
   try {
     const errors = validationResult(req);
@@ -82,4 +59,4 @@ export const removeCard = async (req, res) => {
         message: "Card update error",
       });
     }
-  };
+};
