+++ src/routes/columnRoute.js
@@ -1,0 +1,9 @@
+import * as ColumnController from "../controllers/columnController.js";
+import express from "express";
+
+const router = express.Router();
+
+router.route('/')
+.post(ColumnController.getCardsForColumn)
+
+export default router;
